move breakpoint type
