+++ packages/web/components/chart/asset-breakdown.tsx
@@ -2,8 +2,7 @@ import { CoinPretty, Dec, Int, IntPretty } from "@keplr-wallet/unit";
 import classNames from "classnames";
 import { FunctionComponent } from "react";
 
-import { Breakpoint } from "~/components/types";
-import { useWindowSize } from "~/hooks";
+import { Breakpoint, useWindowSize } from "~/hooks";
 import { truncateString } from "~/utils/string";
 
 const ColorCycle = [
